Reorganize project structure and switch to absolute imports
- switching from code living inside the src/ to src/gensim/ helps python know that everything inside there should be part of a package named gensim
- absolute imports allow the code to be run directly as it was before, but also with python -m gensim.XXX
- remove unused imports and reorganize some imports
